refactor: roleId 타입 변경

+++ techeerzip/src/main/java/backend/techeerzip/domain/user/dto/response/GetUserResponse.java
@@ -1,7 +1,9 @@
 package backend.techeerzip.domain.user.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -35,7 +37,7 @@ public class GetUserResponse {
     private final Long id;
     private final String profileImage;
     private final String name;
-    private final Integer roleId;
+    private final Long roleId;
     private final String nickname;
     private final String email;
     private final String school;
